doc: move trace-gc doc to script
