--- documentation/memory/step3/exercice/.gitignore
@@ -1,1 +1,0 @@
-node_modules
