--- documentation/memory/step3/exercice/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "trace-gc-exercice",
-  "version": "0.0.1",
-  "description": "Exercice for the trace-gc documentation",
-  "main": "server.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "keywords": [
-    "trace-gc",
-    "v8",
-    "node",
-    "diagnostic",
-    "heap"
-  ],
-  "author": "Tony Gorez <gorez.tony@gmail.com>",
-  "license": "MIT",
-  "dependencies": {
-    "fastify": "^4.2.1"
-  }
-}
