+++ documentation/memory/step3/exercice/script.mjs
@@ -1,0 +1,31 @@
+import os from 'os';
+
+let len = 1_000_000;
+const entries = new Set();
+
+function addEntry () {
+  const entry = {
+    timestamp: Date.now(),
+    memory: os.freemem(),
+    totalMemory: os.totalmem(),
+    uptime: os.uptime(),
+  };
+
+  entries.add(entry);
+}
+
+function summary () {
+  console.log(`Total: ${entries.size} entries`);
+}
+
+// execution
+(() => {
+  while (len > 0) {
+    addEntry();
+    process.stdout.write(`~~> ${len} entries to record\r`);
+    len--;
+  };
+
+  summary();
+})();
+
